+++ CHANGES.md
@@ -2,6 +2,7 @@
 # 3.0.0a5 (TBD)
 
 * allow the definition of `geographic_crs` used in the `geographic_bounds` property (https://github.com/cogeotiff/rio-tiler/pull/458)
+* use `contextlib.ExitStack` to better manager opening/closing rasterio dataset (https://github.com/cogeotiff/rio-tiler/pull/459)
 
 # 3.0.0a4 (2021-11-10)